+++ omahoksapp/package.json
@@ -16,6 +16,7 @@
     "react-draggable": "^4.4.5",
     "react-dropzone": "^14.2.3",
     "react-scripts": "5.0.1",
+    "react-tabs": "^6.0.0",
     "react-touch-events": "^3.0.0",
     "react-uuid": "^2.0.0",
     "uuid": "^9.0.0",
