+++ omahoksapp/src/omaLogo.js
@@ -1,5 +1,6 @@
 import logo from './logo.png'
 
+// function to update the sum of points in all of the periods
 const AllPoints = (props) => {
   const summa = []
       Object.entries(props).map(([columnId, column]) => {
@@ -12,12 +13,11 @@ const AllPoints = (props) => {
         })
       })
       return (
-        
-          summa.reduce((acc, currentValue) => acc + currentValue, 0)
-       
-      )
+        summa.reduce((acc, currentValue) => acc + currentValue, 0)
+       )
     }
 
+  
 const MainLogo = (props) => {
     return (
       <div style={{backgroundColor:'#b32d84', position: 'absolute', top:-50, left: -1, right: -1}}>
